Integrate named-axis avals without behavior change
We are making the JAX type system more expressive by adding a
"named shape" to represent SPMD axes, and ending the duck-typing
of ordinary programs and SPMD programs. This is in preparation for
changes that enhance the expressiveness of the JAX frontend, bringing
named/SPMD axes towards parity with positional axes in terms of
features and ergonomics.

This requires many adjustments, including:
- introducing a `pbroadcast` primitive that adds a named axis to a
  value that lacks one
- using `pbroadcast` and named-axis avals throughout the system
  (e.g., for map primitives)
- adding `pbroadcast` to autodiff endpoints to ensure that they
  continue to act "elementwise" over named axes

There are also a few improvements included that are enabled by
these changes:
- removing the implied `pbroadcast` after `psum`, since the
  existing behavior only relies on the `pbroadcast` happening
  eventually (and it will in all existing code)
- revising map primitives to treat return values without the named
  axis as "unmapped"
- taking advantage of this to avoid unnecessary broadcasts/sums
  in partial_eval and transpose of map
- widen the `psum(constant)` special case to support `psum` of
  any value known not to vary across the SPMD axis